Move to utils/download.js
